refactor: clean up src structure
