+++ src/app/routes/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import { Navigate } from 'react-router-dom';
 
-import { AppRoute } from '../shared/config/route';
-import { AuthorizationStatus } from '../shared/config/auth-status';
+import { AppRoute } from '../../shared/config/route';
+import { AuthorizationStatus } from '../../shared/config/auth-status';
 
 type PrivateRouteProps = {
   children: JSX.Element;
