+++ src/entities/PlaceCard/PlaceCard.tsx
