+++ src/shared/config/auth-status.ts
