+++ src/shared/lib/map-utils/map-icons.ts
