+++ src/shared/types/Offer.type.ts
