+++ src/shared/ui/StarsRating/StarsRating.tsx
