+++ src/shared/ui/StarsRating/index.ts
