+++ src/widgets/OfferList/OfferList.tsx
