+++ src/widgets/OfferList/index.ts
